Wont fix : navbar collapsing.
The collapsing navbar does not work and it is too intricate a problem
for me to solve. If you ever want to tackle it :
Hint #1 : Adding the bootstrap example code for a collapsible navbar
textually in the template does not work. May be related to the way bootstrap is used in this
project.
Hint #2 : Some of the elements are missing from the optimised templates.
This does not feel right.

Quick fix : making the navbar non-collapsible.